move sklearn-related tests out of test_export_tree_data.py to sklearn tests folder
